created submit activity

+++ app/src/main/java/com/example/team30app/Coordinates.java
@@ -1,26 +1,26 @@
 package com.example.team30app;
 public class Coordinates {
-    private double longitude;
-    private double latitude;
+    private float longitude;
+    private float latitude;
 
-    public Coordinates(double longitude, double latitude){
+    public Coordinates(float longitude, float latitude){
         this.longitude = longitude;
         this.latitude = latitude;
     }
 
-    public double getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(double longitude) {
+    public void setLongitude(float longitude) {
         this.longitude = longitude;
     }
 
-    public double getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(double latitude) {
+    public void setLatitude(float latitude) {
         this.latitude = latitude;
     }
 }
