+++ app/src/main/java/com/example/team30app/ILocation.java
@@ -2,7 +2,7 @@ package com.example.team30app;
 
 public interface ILocation {
     public Coordinates getCoordinates();
-    public double getLongitude();
-    public double getLatitude();
+    public float getLongitude();
+    public float getLatitude();
     public String getName();
 }
